Added token authentication in saved auctions

+++ src/main/java/com/morse_coders/aucdaisbackend/SavedAuctions/SavedAuctionService.java
@@ -3,11 +3,15 @@ package com.morse_coders.aucdaisbackend.SavedAuctions;
 
 import com.morse_coders.aucdaisbackend.Auction_Products.AuctionProductRepository;
 import com.morse_coders.aucdaisbackend.Auction_Products.AuctionProducts;
+import com.morse_coders.aucdaisbackend.Session.SessionToken;
+import com.morse_coders.aucdaisbackend.Session.SessionTokenRepository;
 import com.morse_coders.aucdaisbackend.Users.Users;
 import com.morse_coders.aucdaisbackend.Users.UsersRepository;
+import org.apache.catalina.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -18,11 +22,14 @@ public class SavedAuctionService {
 
     private final UsersRepository usersRepository;
 
+    private final SessionTokenRepository sessionTokenRepository;
+
     @Autowired
-    public SavedAuctionService(SavedAuctionRepository savedAuctionRepository, AuctionProductRepository auctionProductRepository, UsersRepository usersRepository) {
+    public SavedAuctionService(SavedAuctionRepository savedAuctionRepository, AuctionProductRepository auctionProductRepository, UsersRepository usersRepository, SessionTokenRepository sessionTokenRepository) {
         this.savedAuctionRepository = savedAuctionRepository;
         this.auctionProductRepository = auctionProductRepository;
         this.usersRepository = usersRepository;
+        this.sessionTokenRepository = sessionTokenRepository;
     }
 
     public List<SavedAuctions> getAllSavedAuctions() {
@@ -30,19 +37,32 @@ public class SavedAuctionService {
     }
 
 
-    public void createSavedAuction(SavedAuctions savedAuctions, Long userId, Long auctionId) {
+    public void createSavedAuction(SavedAuctions savedAuctions, Long userId, Long auctionId, String token) {
 
         Optional<Users> user = usersRepository.findById(userId);
         Optional<AuctionProducts> auctionProduct = auctionProductRepository.findById(auctionId);
 
         if (user.isPresent() && auctionProduct.isPresent()) {
-            // check if same user saved the same product, if yes, return error
-            if (savedAuctionRepository.findByUserIdAndAuctionId(userId, auctionId) != null) {
-                throw new IllegalArgumentException("User already saved this product");
+            Optional<SessionToken> getUserToken = sessionTokenRepository.findByUserAndExpiresAt(user.get(), LocalDateTime.now());
+            if(getUserToken.isPresent()){
+                if(getUserToken.get().getToken().equals(token)){
+                    // check if same user saved the same product, if yes, return error
+                    if (savedAuctionRepository.findByUserIdAndAuctionId(userId, auctionId) != null) {
+                        throw new IllegalArgumentException("User already saved this product");
+                    }
+                    savedAuctions.setUser(user.get());
+                    savedAuctions.setAuctionProduct(auctionProduct.get());
+                    savedAuctionRepository.save(savedAuctions);
+                }
+                else {
+                    throw new RuntimeException("Token is expired / not valid");
+                }
+
             }
-            savedAuctions.setUser(user.get());
-            savedAuctions.setAuctionProduct(auctionProduct.get());
-            savedAuctionRepository.save(savedAuctions);
+            else {
+                throw new RuntimeException("Token is expired / not valid");
+            }
+
         }
         else{
             System.out.println("User or AuctionProduct not found");
@@ -61,8 +81,28 @@ public class SavedAuctionService {
     }
 
 
-    public List<SavedAuctions> getAllSavedAuctionsOfUser(Long userId) {
-        return savedAuctionRepository.findAllByUserId(userId);
+    public List<SavedAuctions> getAllSavedAuctionsOfUser(Long userId, String token) {
+        Optional<Users> user = usersRepository.findById(userId);
+        if(user.isPresent()){
+            Optional<SessionToken> getUserToken = sessionTokenRepository.findByUserAndExpiresAt(user.get(), LocalDateTime.now());
+            if(getUserToken.isPresent()){
+                if(getUserToken.get().getToken().equals(token)){
+                    return savedAuctionRepository.findAllByUserId(userId);
+                }
+                else {
+                    throw new RuntimeException("Token is expired / not valid");
+                }
+
+            }
+            else {
+                throw new RuntimeException("Token is expired / not valid");
+            }
+
+        }
+        else{
+            System.out.println("User not found");
+        }
+        return null;
     }
 
     // all saved auction by user id and auction id
@@ -71,20 +111,69 @@ public class SavedAuctionService {
     }
 
     // all saved auction by user id and auction id before a specific date
-    public List<SavedAuctions> getAllSavedAuctionsByUserIdAndAuctionIdBeforeDate(Long userId, String date) {
-        return savedAuctionRepository.findAllByUserIdAndAuctionProductIdBeforeDate(userId, date);
+    public List<SavedAuctions> getAllSavedAuctionsByUserIdAndAuctionIdBeforeDate(Long userId, String date, String token) {
+        Optional<Users> user = usersRepository.findById(userId);
+        if(user.isPresent()){
+            Optional<SessionToken> getUserToken = sessionTokenRepository.findByUserAndExpiresAt(user.get(), LocalDateTime.now());
+            if(getUserToken.isPresent()){
+                if(getUserToken.get().getToken().equals(token)){
+                    return savedAuctionRepository.findAllByUserIdAndAuctionProductIdBeforeDate(userId, date);
+                }
+                else {
+                    throw new RuntimeException("Token is expired / not valid");
+                }
+            }
+            else {
+                throw new RuntimeException("Token is expired / not valid");
+            }
+        }
+        else{
+            System.out.println("User not found");
+        }
+        return null;
     }
 
     // all saved auction by user id and auction id after a specific date
-    public List<SavedAuctions> getAllSavedAuctionsByUserIdAndAuctionIdAfterDate(Long userId, String date) {
-        return savedAuctionRepository.findAllByUserIdAndAuctionProductIdAfterDate(userId, date);
+    public List<SavedAuctions> getAllSavedAuctionsByUserIdAndAuctionIdAfterDate(Long userId, String date, String token) {
+        Optional<Users> user = usersRepository.findById(userId);
+        if (user.isPresent()){
+            Optional<SessionToken> getUserToken = sessionTokenRepository.findByUserAndExpiresAt(user.get(), LocalDateTime.now());
+            if(getUserToken.isPresent()){
+                if(getUserToken.get().getToken().equals(token)){
+                    return savedAuctionRepository.findAllByUserIdAndAuctionProductIdAfterDate(userId, date);
+                }
+                else {
+                    throw new RuntimeException("Token is expired / not valid");
+                }
+            }
+            else {
+                throw new RuntimeException("Token is expired / not valid");
+            }
+        }
+        else{
+            System.out.println("User not found");
+        }
+        return null;
     }
 
-    public void deleteSavedAuctionByUserIdAndAuctionId(Long userId, Long auctionId) {
+    public void deleteSavedAuctionByUserIdAndAuctionId(Long userId, Long auctionId, String token) {
         Optional<Users> users = usersRepository.findById(userId);
         Optional<AuctionProducts> auctionProduct = auctionProductRepository.findById(auctionId);
         if (users.isPresent() && auctionProduct.isPresent()) {
-            savedAuctionRepository.deleteSavedAuctionByUserIdAndAuctionProductId(userId, auctionId);
+            Optional<SessionToken> getUserToken = sessionTokenRepository.findByUserAndExpiresAt(users.get(), LocalDateTime.now());
+            if(getUserToken.isPresent()){
+                if(getUserToken.get().getToken().equals(token)){
+                    savedAuctionRepository.deleteSavedAuctionByUserIdAndAuctionProductId(userId, auctionId);
+                }
+                else {
+                    throw new RuntimeException("Token is expired / not valid");
+                }
+
+            }
+            else {
+                throw new RuntimeException("Token is expired / not valid");
+            }
+
         }
         else{
             System.out.println("User or AuctionProduct not found");
